add explicit root object This adds a explicit root object, which is used as the source object for the root query instead of context.
Adds also a new ExecutionInput to be used as Input for GraphQL.execute.

Important: Old GraphQL.execute calls are unchanged: root is set to context
in order to not break it!

other breaking changes (just renaming) made along the way to make the overall
naming better and to avoid confusion with the new ExecutionInput

- instrumentation.ExecutionParameters is renamed to InstrumentationExecutionParameters
- instrumentation.ValidationParameters is renamed to InstrumentationValidationParameters
- ExecutionParameters is renamed to ExecutionStrategyParameters